improvement(compiler) More x86_64 speicific code removed from CodeGen
